feat(blog): persist sidebar shell across post navigation
